test(sessions): keep idempotency fixture at test boundary
